fix(gv-code): resolve dispatch of change event
fix gravitee-io/issues#4712
