+++ src/lib/studio.js
@@ -124,11 +124,3 @@ export function getFlowName (flow, collectionName, withMethods = true, draggable
   }
   return rendering;
 }
-
-export function uuid () {
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-    const r = Math.random() * 16 | 0;
-    const v = c === 'x' ? r : (r & 0x3 | 0x8);
-    return v.toString(16);
-  });
-}
