+++ stories/resources/schemas/groovy.json
@@ -1,0 +1,70 @@
+{
+  "type" : "object",
+  "id" : "urn:jsonschema:io:gravitee:policy:groovy:configuration:GroovyPolicyConfiguration",
+  "properties" : {
+    "onRequestScript" : {
+      "title": "On-request script",
+      "description": "Groovy script to evaluate during the OnRequest phase.",
+      "type" : "string",
+      "x-schema-form": {
+        "type": "codemirror",
+        "codemirrorOptions": {
+          "placeholder": "Put your groovy script here or drag'n'drop groovy script",
+          "lineWrapping": true,
+          "lineNumbers": true,
+          "allowDropFileTypes": true,
+          "autoCloseTags": true,
+          "mode": "groovy"
+        }
+      }
+    },
+    "onResponseScript" : {
+      "title": "On-response script",
+      "description": "Groovy script to evaluate during the OnResponse phase.",
+      "type" : "string",
+      "x-schema-form": {
+        "type": "codemirror",
+        "codemirrorOptions": {
+          "placeholder": "Put your groovy script here or drag'n'drop groovy script",
+          "lineWrapping": true,
+          "lineNumbers": true,
+          "allowDropFileTypes": true,
+          "autoCloseTags": true,
+          "mode": "groovy"
+        }
+      }
+    },
+    "onRequestContentScript" : {
+      "title": "On-request content script",
+      "description": "Groovy script to evaluate during the 'request-content' phase. <br> Useful to transform request content",
+      "type" : "string",
+      "x-schema-form": {
+        "type": "codemirror",
+        "codemirrorOptions": {
+          "placeholder": "Put your groovy script here or drag'n'drop groovy script",
+          "lineWrapping": true,
+          "lineNumbers": true,
+          "allowDropFileTypes": true,
+          "autoCloseTags": true,
+          "mode": "groovy"
+        }
+      }
+    },
+    "onResponseContentScript" : {
+      "title": "On-response content script",
+      "description": "Groovy script to evaluate during the 'response-content' phase. <br> Useful to transform response content",
+      "type" : "string",
+      "x-schema-form": {
+        "type": "codemirror",
+        "codemirrorOptions": {
+          "placeholder": "Put your groovy script here or drag'n'drop groovy script",
+          "lineWrapping": true,
+          "lineNumbers": true,
+          "allowDropFileTypes": true,
+          "autoCloseTags": true,
+          "mode": "groovy"
+        }
+      }
+    }
+  }
+}
